+++ scene/resources/style_box.cpp
@@ -136,7 +136,7 @@ Ref<Texture> StyleBoxTexture::get_normal_map() const {
 
 void StyleBoxTexture::set_margin_size(Margin p_margin, float p_size) {
 
-	ERR_FAIL_INDEX(p_margin, 4);
+	ERR_FAIL_INDEX((int)p_margin, 4);
 
 	margin[p_margin] = p_size;
 	emit_changed();
@@ -200,7 +200,7 @@ Size2 StyleBoxTexture::get_center_size() const {
 
 void StyleBoxTexture::set_expand_margin_size(Margin p_expand_margin, float p_size) {
 
-	ERR_FAIL_INDEX(p_expand_margin, 4);
+	ERR_FAIL_INDEX((int)p_expand_margin, 4);
 	expand_margin[p_expand_margin] = p_size;
 	emit_changed();
 }
@@ -223,7 +223,7 @@ void StyleBoxTexture::set_expand_margin_size_all(float p_expand_margin_size) {
 
 float StyleBoxTexture::get_expand_margin_size(Margin p_expand_margin) const {
 
-	ERR_FAIL_INDEX_V(p_expand_margin, 4, 0);
+	ERR_FAIL_INDEX_V((int)p_expand_margin, 4, 0);
 	return expand_margin[p_expand_margin];
 }
